11.1-Z.3 seal: allocator-UB fix + fail-closed oracle-isolated signature court + safety proof scopes (R-000178)
- default allocator is now plain libc malloc/realloc/free/strdup (upstream
  2.15.0 globals.c defaults), eliminating the invalid-layout UB (Rust
  std::alloc with fabricated Layouts); the debug-named surface
  (xmlMemMalloc/*Loc) keeps the tracked registry; xmlMemUsed/xmlMemBlocks/
  xmlMemSize and the display entry points match the oracle byte-for-byte
  (R-000178, ALLOCATOR-DEFAULT-001 differential court)
- ABI-FUNCTION-SIGNATURE hardened: fully separated oracle/candidate include
  environments with per-include-origin contamination checks, fail-closed
  header/fn accounting (silent_omissions=0), and two fingerprints
  (SOURCE_PROTOTYPE + MACHINE_ABI: signedness, pointee identity, full
  fn-pointer signatures, return pointer depth) — found and fixed 25+ real
  divergences (xmlBufferGrow/Shrink unsigned, xmlDict*/xmlOutputBufferGetSize
  size_t, xmlValidateAttributeDecl 4->3 args, xmlXPathCastBooleanToNumber
  double, xmlCheckVersion/xmlInitThreads void, xsltPreComputeFunction/
  xsltTopLevelFunction typedefs, removed invented xmlTextReaderGetAttributeIndex,
  char/xmlChar pointee drifts, xmlAttributeTable typedef identities, ...)
- CUSTODIAN-COMMENTARY-DRIFT now verdict-gates on zero unaccounted safety
  sites via proof scopes (SAFETY-SCOPE markers + fn-doc safety contracts):
  unsafe fns 1811/1811, unsafe blocks 5235/5235 (local 2514 + enclosing-scope
  2720 + generated 1); rustdoc gate (RUSTDOCFLAGS=-D warnings cargo doc)
  added and clean
- xslHandleDebugger reclassified HEADER_DECLARED_ORACLE_DSO_ABSENT (the
  oracle DSO itself does not export it); xsltCleanupSourceDoc likewise
- headers: 30+ candidate-header declarations aligned to the oracle text;
  R-000131 narrative sealed to the Z.3 state; R-000177 framed as the Phase-12
  DSO-boundary-lint architectural target
- ledger: 82 residuals, 79 FIXED / 3 OPEN
- version 0.1.0-alpha.36

+++ include/libxml/dict.h
@@ -7,6 +7,7 @@
 #ifndef __XML_DICT_H__
 #define __XML_DICT_H__
 
+#include <stddef.h>
 #include <libxml/xmlversion.h>
 #include <libxml/xmlstring.h>
 
@@ -21,10 +22,10 @@ XMLPUBFUN xmlDictPtr xmlDictCreate(void);
 XMLPUBFUN xmlDictPtr xmlDictCreateSub(xmlDictPtr sub);
 XMLPUBFUN const xmlChar *xmlDictLookup(xmlDictPtr dict, const xmlChar *name, int len);
 XMLPUBFUN const xmlChar *xmlDictExists(xmlDictPtr dict, const xmlChar *name, int len);
-XMLPUBFUN unsigned int xmlDictSize(const xmlDictPtr dict);
+XMLPUBFUN int xmlDictSize(const xmlDictPtr dict);
 XMLPUBFUN void xmlDictFree(xmlDictPtr dict);
-XMLPUBFUN unsigned int xmlDictSetLimit(xmlDictPtr dict, unsigned int limit);
-XMLPUBFUN unsigned int xmlDictGetUsage(const xmlDictPtr dict);
+XMLPUBFUN size_t xmlDictSetLimit(xmlDictPtr dict, size_t limit);
+XMLPUBFUN size_t xmlDictGetUsage(const xmlDictPtr dict);
 
 
 /* [11.1-S] begin: oracle-extracted declarations
